bot/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-telegram-bot==20.6
python-dotenv==1.0.1
